+++ apps/client/package.json
@@ -34,6 +34,8 @@
     "web-vitals": "^3.1.1"
   },
   "scripts": {
+    "addversion": "node -p \"'export const ONTIME_VERSION = ' + JSON.stringify(require('../../package.json').version) + ';'\" > src/ONTIME_VERSION.js",
+    "postinstall": "pnpm addversion",
     "dev": "set BROWSER=none&&vite",
     "build": "vite build",
     "lint": "eslint .",
