+++ apps/electron/package.json
@@ -19,9 +19,8 @@
     "prettier": "^2.8.3"
   },
   "scripts": {
-    "addversion": "node -p \"'export const ONTIME_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/version.mjs",
-    "postinstall": "pnpm addversion",
-    "dev": "NODE_ENV=development electron .",
+    "postinstall": "",
+    "dev:electron": "NODE_ENV=development electron .",
     "dist-win": "electron-builder --publish=never  --x64 --win",
     "dist-mac": "electron-builder --publish=never  --x64 --mac",
     "dist-linux": "electron-builder --publish=never  --x64 --linux",
