--- apps/server/.npmrc
@@ -1,1 +1,0 @@
-node-linker=hoisted
