+++ apps/server/package.json
@@ -16,14 +16,12 @@
     "lowdb": "^5.0.5",
     "multer": "^1.4.4",
     "nanoid": "^4.0.0",
-    "node-osc": "^8.0.6",
+    "node-osc": "^8.0.9",
     "node-xlsx": "^0.21.0",
     "ontime-utils": "workspace:*",
     "passport": "^0.6.0",
     "passport-local": "~1.0.0",
-    "socket.io": "^4.5.4",
-    "ts-node-dev": "^2.0.0",
-    "tsup": "^6.5.0"
+    "socket.io": "^4.5.4"
   },
   "devDependencies": {
     "@types/express": "^4.17.15",
@@ -35,29 +33,24 @@
     "eslint-plugin-prettier": "^4.2.1",
     "nodemon": "^2.0.20",
     "prettier": "^2.8.3",
+    "shx": "^0.3.4",
     "ts-node": "^10.9.1",
     "typescript": "^4.9.4",
     "vitest": "^0.27.1"
   },
   "scripts": {
+    "addversion": "node -p \"'export const ONTIME_VERSION = ' + JSON.stringify(require('../../package.json').version) + ';'\" > src/ONTIME_VERSION.js",
+    "postinstall": "pnpm addversion",
     "dev": "set NODE_ENV=development && nodemon --exec \"ts-node-esm\" ./src/index.ts",
     "dev:test": "set IS_TEST=true && nodemon --exec \"ts-node-esm\" ./src/index.ts",
-    "build": "esbuild src/app.ts --platform=node --format=cjs --bundle --minify --outfile=dist/index.cjs",
-    "build:debug": "esbuild src/app.ts --platform=node --format=cjs --bundle --outfile=dist/index.cjs",
-    "postinstall": "pnpm setdb",
+    "setdb": "shx cp ../../demo-db/db.json src/preloaded-db/db.json",
+    "prebuild": "pnpm setdb",
+    "build": "pnpm prebuild && esbuild src/app.ts --log-level=error --platform=node --format=cjs --bundle --minify --outfile=dist/index.cjs",
+    "build:debug": "pnpm prebuild && esbuild src/app.ts --platform=node --format=cjs --bundle --outfile=dist/index.cjs",
     "lint": "eslint .",
     "test": "vitest",
     "test:pipeline": "vitest run",
     "typecheck": "tsc --noEmit",
     "cleanup": "rm -rf .turbo && rm -rf node_modules && rm -rf dist"
-  },
-  "tsup": {
-    "entry": [
-      "src/index.ts"
-    ],
-    "sourcemap": true,
-    "clean": true,
-    "shims": true,
-    "minify": true
   }
 }
