--- apps/server/preloaded-db/db.json
@@ -1,150 +1,0 @@
-{
-  "rundown": [
-    {
-      "title": "Welcome to Ontime",
-      "subtitle": "Subtitles are useful",
-      "presenter": "cpvalente",
-      "note": "Maybe a running note for the operator?",
-      "timeStart": 28800000,
-      "timeEnd": 30600000,
-      "isPublic": false,
-      "type": "event",
-      "revision": 7,
-      "id": "5946"
-    },
-    {
-      "title": "This is your event list",
-      "subtitle": "",
-      "presenter": "cpvalente",
-      "note": "",
-      "timeStart": 30600000,
-      "timeEnd": 34200000,
-      "isPublic": false,
-      "type": "event",
-      "revision": 4,
-      "id": "c2e7"
-    },
-    {
-      "title": "Events run in the list order",
-      "subtitle": "",
-      "presenter": "",
-      "note": "",
-      "timeStart": 34200000,
-      "timeEnd": 34800000,
-      "isPublic": false,
-      "type": "event",
-      "revision": 4,
-      "id": "cc0f"
-    },
-    {
-      "title": "Unless recalled by the OSC address",
-      "subtitle": "",
-      "presenter": "",
-      "note": "In green, below",
-      "timeStart": 34800000,
-      "timeEnd": 35400000,
-      "isPublic": false,
-      "type": "event",
-      "revision": 7,
-      "id": "8ee5"
-    },
-    {
-      "title": "Use simpler times to create a timer",
-      "subtitle": "",
-      "presenter": "",
-      "note": "",
-      "timeStart": 0,
-      "timeEnd": 600000,
-      "isPublic": false,
-      "type": "event",
-      "revision": 2,
-      "id": "8222"
-    },
-    {
-      "duration": 900000,
-      "type": "delay",
-      "revision": 1,
-      "id": "a386"
-    },
-    {
-      "title": "Add delay blocks to affect all events",
-      "subtitle": "",
-      "presenter": "",
-      "note": "* Until a block is found",
-      "timeStart": 35400000,
-      "timeEnd": 36600000,
-      "isPublic": false,
-      "type": "event",
-      "revision": 6,
-      "id": "6dce"
-    },
-    {
-      "title": "Add and remove events with [+] and [-]",
-      "subtitle": "",
-      "presenter": "",
-      "note": "Orange and red buttons on the right",
-      "timeStart": 36600000,
-      "timeEnd": 43200000,
-      "isPublic": false,
-      "type": "event",
-      "revision": 9,
-      "id": "2651"
-    },
-    {
-      "type": "block",
-      "id": "e6a1"
-    },
-    {
-      "title": "And control whether they are public",
-      "subtitle": "",
-      "presenter": "",
-      "note": "Blue button on the right",
-      "timeStart": 46800000,
-      "timeEnd": 57600000,
-      "isPublic": false,
-      "type": "event",
-      "revision": 6,
-      "id": "1358"
-    }
-  ],
-  "event": {
-    "title": "All about Carlos demo event",
-    "url": "www.carlosvalente.com",
-    "publicInfo": "WiFi: demoproject \nPassword: ontimeproject",
-    "backstageInfo": "WiFi: demobackstage\nPassword: ontimeproject",
-    "endMessage": ""
-  },
-  "settings": {
-    "app": "ontime",
-    "version": 2,
-    "serverPort": 4001,
-    "lock": null,
-    "pinCode": "1234"
-  },
-  "osc": {
-    "port": 8888,
-    "portOut": 9999,
-    "targetIP": "127.0.0.1",
-    "enabled": true
-  },
-  "aliases": [
-    {
-      "id": "0b0b3",
-      "enabled": true,
-      "alias": "test",
-      "pathAndParams": "lower?bg=ff2&text=f00&size=0.6&transition=5"
-    }
-  ],
-  "userFields": {
-    "user0": "user0",
-    "user1": "user1",
-    "user2": "user2",
-    "user3": "user3",
-    "user4": "user4",
-    "user5": "user5",
-    "user6": "user6",
-    "user7": "user7",
-    "user8": "user8",
-    "user9": "user9"
-  }
-}
