+++ apps/server/src/modules/loadDb.js
@@ -6,15 +6,15 @@ import { validateFile } from '../utils/parserUtils.js';
 import { dbModel } from '../models/dataModel.js';
 import { parseJson } from '../utils/parser.js';
 import { reportSentryException } from './sentry.js';
-import { currentDirectory, pathToStartDb, resolveDbPath } from '../setup.js';
+import { pathToStartDb, resolveDbDirectory, resolveDbPath } from '../setup.js';
 
 /**
  * @description ensures directories exist and populates database
  * @return {string} - path to db file
  */
 const populateDb = () => {
-  const dbInDisk = resolveDbPath();
-  ensureDirectory(dbInDisk);
+  const dbInDisk = resolveDbPath;
+  ensureDirectory(resolveDbDirectory);
 
   // if dbInDisk doesn't exist we want to use startup db
   if (!existsSync(dbInDisk)) {
@@ -45,7 +45,7 @@ const parseDb = async (fileToRead, adapterToUse) => {
 };
 
 /**
- * @description Modules loads ontime db
+ * @description loads ontime db
  * @return {Promise<{data: (*), db: Low<unknown>}>}
  */
 async function loadDb() {
