--- utils/src/generate-id/generateId.js
@@ -1,3 +1,0 @@
-import { customAlphabet } from 'nanoid';
-const nanoid = customAlphabet('1234567890abcdef', 5);
-export const generateId = () => nanoid();
